+++ fresh.gen.ts
@@ -10,6 +10,7 @@ import * as $api_db_tournaments_create from "./routes/api/db/tournaments/create.
 import * as $api_slippi_cleanDirectory from "./routes/api/slippi/cleanDirectory.ts";
 import * as $api_slippi_getCurrentGame from "./routes/api/slippi/getCurrentGame.ts";
 import * as $api_startgg_searchTournaments from "./routes/api/startgg/searchTournaments.ts";
+import * as $api_startgg_startSet from "./routes/api/startgg/startSet.ts";
 import * as $api_startgg_updateSet from "./routes/api/startgg/updateSet.ts";
 import * as $index from "./routes/index.tsx";
 import * as $set_setId_ from "./routes/set/[setId].tsx";
@@ -19,6 +20,7 @@ import * as $tournament_tournamentSlug_index from "./routes/tournament/[tourname
 import * as $InteractiveSet from "./islands/InteractiveSet.tsx";
 import * as $PlayerInterface from "./islands/PlayerInterface.tsx";
 import * as $PortSelector from "./islands/PortSelector.tsx";
+import * as $StartNextMatchButton from "./islands/StartNextMatchButton.tsx";
 import * as $TournamentSearch from "./islands/TournamentSearch.tsx";
 import { type Manifest } from "$fresh/server.ts";
 
@@ -32,6 +34,7 @@ const manifest = {
     "./routes/api/slippi/cleanDirectory.ts": $api_slippi_cleanDirectory,
     "./routes/api/slippi/getCurrentGame.ts": $api_slippi_getCurrentGame,
     "./routes/api/startgg/searchTournaments.ts": $api_startgg_searchTournaments,
+    "./routes/api/startgg/startSet.ts": $api_startgg_startSet,
     "./routes/api/startgg/updateSet.ts": $api_startgg_updateSet,
     "./routes/index.tsx": $index,
     "./routes/set/[setId].tsx": $set_setId_,
@@ -44,6 +47,7 @@ const manifest = {
     "./islands/InteractiveSet.tsx": $InteractiveSet,
     "./islands/PlayerInterface.tsx": $PlayerInterface,
     "./islands/PortSelector.tsx": $PortSelector,
+    "./islands/StartNextMatchButton.tsx": $StartNextMatchButton,
     "./islands/TournamentSearch.tsx": $TournamentSearch,
   },
   baseUrl: import.meta.url,